[disk] Return the actual device instead of /dev/root
If psutil.disk_partitions() might return '/dev/root' on some Linux setups.
When that happens, the new code finds the actual device it maps to.
Otherwise, when we later call `blkid` to look for disk labels, we would
find no labels for the root disk as `blkid` prints the actual device path
instead of /dev/root