Clarify the instructions releasing integration (#11906)
* Clarify the instructions releasing integration

Step 2 was a bit unclear. The instructions made it seem like the integration release pipeline triggering can be done in parallel with the creation of the PR. I moved it to its own step to hopefully make it clearer that the release pipeline should be triggered after the PR has been merged in.

* Doc Review

Co-authored-by: Sarah Witt <sarah.witt@datadoghq.com>

Co-authored-by: Austin Lai <76412946+alai97@users.noreply.github.com>
Co-authored-by: Sarah Witt <sarah.witt@datadoghq.com> ef23527aa785062fe976add0d939f124c3d50b56