Fix incorrect OpenMetrics V2 check exposition format HTTP header (#11899)
* Fix incorrect exposition format header with OpenMetricsV2 check

The exposition format to request is set via the "Accept" header. Despite
the OpenMetricsScraper appearing to set the correct header when
`use_latest_spec` is enabled, it is actually a no-op due to the default
HTTP headers including `Accept: */*` and the use of `setdefault`.

With this change, the openmetrics check should request the appropriate
exposition format fixing "Clashing name" errors.

* Add tests d979458dcc4e120b11bd4a2fffd7abe65aa006a4