Check for missing assets on validation (saved views and config) (#12359)
Saved views validations didn't report failures when a file in the
manifest is not found, as opposed to most of the other asset
validation, and this fixes that.

This also adds a warning when the config spec is not found. fe2f19d85cca8a0a32ec5568a4faf2c56ee24d76