Tidy up Vault integration's metadata.csv (#12355)
* Update vault.route metadata to reflect OpenMetricsV2 as default

This includes:

- Marking all metrics that include the mountpoint (whatever comes
after `vault.route.<api>.`) as legacy because those are not emitted
separately with OpenMetricsV2.
- For the rest, the metrics with a stripped `vault.` prefix in the
metrics are marked as legacy and new ones with the double prefix have
been added as necessary, reflecting the fact that we don't strip the
prefix in the V2 implementation.
- A few minor corrections in descriptions and such.

* Revert legacy sum metrics to gauge type in metadata.csv

This metrics are only sent when using the legacy v1 implementation,
which sends them as gauges. 45a159c0eb07e1ef75e9d8ff7dd24fb268b079ff